docker/genie-requirements.txt
pyyaml
nipyapi
minio